terakan: Use RTV/UAV terms instead of MRT/RAT
AMD uses MRT to refer to color targets and RATs collectively, use that
abbreviation the same way in Terakan too, but refer to color targets
specifically as RTVs (the term "color attachment" is also not precise
enough in Terakan because it refers to color attachments at pre-compaction
indices).

Also perform deratization using an unmanned aerial vehicle because RTV and
RAT sound relatively close to each other.

Signed-off-by: Vitaliy Triang3l Kuzmin <triang3l@yandex.ru>